+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 - [fix: fix RouterLabelRestTemplateInterceptor add response headers exception with httpclient5.](https://github.com/Tencent/spring-cloud-tencent/pull/1267)
 - [feat: support lossless online and offline](https://github.com/Tencent/spring-cloud-tencent/pull/1268)
 - [feat: support lane router](https://github.com/Tencent/spring-cloud-tencent/pull/1269)
+- [feat: add lane router examples](https://github.com/Tencent/spring-cloud-tencent/pull/1270)